requirements.txt
Django==4.0
Pillow==8.4.0
djangorestframework==3.13.0
drf-spectacular==0.21.0
djangorestframework-simplejwt==5.0.0